% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{approve_release}
\alias{approve_release}
\title{Approve a GitHub release}
\usage{
approve_release(pkg = ".", filename = "NEWS.md", host = getOption("rdev.host"))
}
\arguments{
\item{pkg}{path to package. Currently, only \code{pkg = "."} is supported.}

\item{filename}{name of file containing release notes, defaults to \code{NEWS.md}.}

\item{host}{GitHub host to target, passed to the \code{.api_url} argument of
\code{\link[gh:gh]{gh::gh()}}. If unspecified, gh defaults to "https://api.github.com",
although gh's default can be customised by setting the GITHUB_API_URL
environment variable.

For a hypothetical GitHub Enterprise instance, either
"https://github.acme.com/api/v3" or "https://github.acme.com" is
acceptable.}
}
\value{
Staged GitHub pull request, invisibly
}
\description{
Approve a pull request staged with \code{\link[=stage_release]{stage_release()}}, and prepare for release. Merge an approved
release with \code{\link[=merge_release]{merge_release()}}.
}
\details{
When run, \code{approve_release()}:
\enumerate{
\item Checks for uncommitted changes and stops if any exist using \code{\link[gert:git_status]{gert::git_status()}}
\item Determines the staged release title from \code{NEWS.md} using \code{\link[=get_release]{get_release()}}
\item Selects the GitHub pull request that matches the staged release title, stops if there is more
or less than one matching PR using \code{\link[gh:gh]{gh::gh()}}
\item Checks if on the PR branch and stops if not (using \code{\link[gert:git_branch]{gert::git_branch()}})
\item Verifies the release is not already approved by searching for the release commit
\item Updates the PR description if the release notes have changed
\item Updates \code{Version} in \code{DESCRIPTION} with \code{\link[desc:desc_set_version]{desc::desc_set_version()}}, commits and push to git
with message \code{"GitHub release <version>"} using \code{\link[gert:git_add]{gert::git_add()}}, \code{\link[gert:git_commit]{gert::git_commit()}} and
\code{\link[gert:git_push]{gert::git_push()}}
\item Runs \code{\link[=build_quarto_site]{build_quarto_site()}} (if \verb{_quarto.yml} exists), \code{\link[=build_analysis_site]{build_analysis_site()}} (if
\verb{pkgdown/_base.yml} exists) or \code{\link[=build_rdev_site]{build_rdev_site()}} (if \verb{_pkgdown.yml} exists), commits and
pushes changes (if any) to git with message: \code{"<builder> for release <version>"}
}
}
\section{Host}{

Set the \code{rdev.host} option when using a GitHub Enterprise server:
\code{options(rdev.host = "https://github.example.com/api/v3")}
}

