% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\docType{methods}
\name{as.list,mrgmod-method}
\alias{as.list,mrgmod-method}
\title{Coerce an \code{mrgmod} object to list.}
\usage{
\S4method{as.list}{mrgmod}(x, ...)
}
\arguments{
\item{x}{mrgmod object}

\item{...}{passed to other methods}
}
\description{
Coerce an \code{mrgmod} object to list.
}

