% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockedmod.R
\docType{methods}
\name{as.locked}
\alias{as.locked}
\alias{as.locked,mrgmod-method}
\title{Corece an mrgmod object to a lockedmod object.}
\usage{
as.locked(x, ...)

\S4method{as.locked}{mrgmod}(x, dllloc, dllname, src, include, ...)
}
\arguments{
\item{x}{mrgmod model object}

\item{...}{passed along}

\item{dllloc}{directory location for the model shared object}

\item{dllname}{the name of the model shared object}

\item{src}{directory location of the model specification file}

\item{include}{directory location for the header file}
}
\description{
Corece an mrgmod object to a lockedmod object.
}

