% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.matrix.list}
\alias{as.matrix.list}
\title{Coerce a list to a matrix}
\usage{
as.matrix.list(x, ..., nrow = 1)
}
\arguments{
\item{x}{a named list, with each element of length 1}

\item{...}{not used}

\item{nrow}{the number of rows to make the matrix}
}
\value{
matrix with colnames set to the names of x
}
\description{
All elements of the list must be of length 1.
}
\examples{
x <- list(a=1, b=2, c=3)
as.matrix(x,nrow=3)
}
\author{
Kyle Baron
}

