% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\docType{methods}
\name{as.packmod}
\alias{as.packmod}
\alias{as.packmod,mrgmod-method}
\title{Coerce an mrgmod object to packmod}
\usage{
as.packmod(x, ...)

\S4method{as.packmod}{mrgmod}(x, ...)
}
\arguments{
\item{x}{mrgmod model object}

\item{...}{passed along}
}
\description{
Coerce an mrgmod object to packmod
}

