% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_analysis_site}
\alias{build_analysis_site}
\title{Build Analysis Site}
\usage{
build_analysis_site(pkg = ".", ...)
}
\arguments{
\item{pkg}{Path to package. Currently, only \code{pkg = "."} is supported.}

\item{...}{additional arguments passed to \code{\link[pkgdown:build_site_github_pages]{pkgdown::build_site_github_pages()}} (not implemented)}
}
\value{
rmarkdown _site.yml as yaml, invisibly
}
\description{
\code{build_analysis_site()} is a wrapper for \code{\link[pkgdown:build_site_github_pages]{pkgdown::build_site_github_pages()}} that adds an
'Analysis' menu containing rendered versions of all .Rmd files in \verb{analysis/}.
}
\details{
When run, \code{build_analysis_site()}:
\enumerate{
\item Reads base \link{pkgdown} settings from \verb{pkgdown/_base.yml}
\item Writes base settings to \verb{_pkgdown.yml}
\item Creates a template using \code{\link[pkgdown:templates]{pkgdown::template_navbar()}} and inserts an \code{analysis} menu with
links to html versions of each .Rmd file in \verb{analysis/}
\item Writes the template to \verb{_pkgdown.yml}
\item Updates \code{README.md} by running \code{\link[devtools:build_rmd]{devtools::build_readme()}} (if \code{README.Rmd} exists) to update
the list of notebooks
\item Runs \code{\link[pkgdown:build_site_github_pages]{pkgdown::build_site_github_pages()}} with \code{install = TRUE} and \code{new_process = TRUE}
\item Creates a \verb{_site.yml} file based on the final \verb{_pkgdown.yml} that clones the \link{pkgdown} navbar
in a temporary build directory
\item Copies the following from \verb{analysis/} into the build directory: \verb{*.Rmd}, \verb{assets/}, \verb{data/},
\verb{import/}, \verb{rendered/}
\item Changes \verb{*.Rmd} from \code{html_notebook} to \code{html_document} using \code{\link[=to_document]{to_document()}}
\item Builds a site using \code{\link[rmarkdown:render_site]{rmarkdown::render_site()}} using modified \code{html_document} output settings
to render files with the look and feel of \code{html_notebook}
\item Moves the rendered files to \verb{docs/}: \verb{*.html}, \verb{assets/}, \verb{rendered/}, without overwriting
}

\code{build_analysis_site()} will fail with an error if there are no files in \verb{analysis/*.Rmd}, or if
\verb{pkgdown/_base.yml} does not exist.
}
\section{Continuous Integration}{

Both \code{\link[=build_rdev_site]{build_rdev_site()}} and \code{\link[=build_analysis_site]{build_analysis_site()}} are meant to be used as part of a CI/CD
workflow, and temporarily set the environment variable \code{CI == "TRUE"} so that the build will
fail when non-internal topics are not included on the reference index page per
\code{\link[pkgdown:build_reference]{pkgdown::build_reference()}}.
}

\section{Supported File Types}{

While \code{\link[=build_quarto_site]{build_quarto_site()}} supports both R Markdown (\code{.Rmd}) and Quarto (\code{.qmd}) notebooks in
the \code{analysis} directory interchangeably, \code{\link[=build_analysis_site]{build_analysis_site()}} supports \code{.Rmd} files only.
}

