% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_quarto_site}
\alias{build_quarto_site}
\title{Build Quarto Site}
\usage{
build_quarto_site(input = NULL, as_job = FALSE, unfreeze = FALSE, ...)
}
\arguments{
\item{input}{The input file or project directory to be rendered (defaults
to rendering the project in the current working directory).}

\item{as_job}{Render as an RStudio background job. Default is "auto",
which will render individual documents normally and projects as
background jobs. Use the \code{quarto.render_as_job} R option to control
the default globally.}

\item{unfreeze}{If \code{TRUE}, delete the Quarto \verb{_freeze} directory to fully re-render the site.}

\item{...}{Arguments passed to \code{\link[quarto:quarto_render]{quarto::quarto_render()}}.}
}
\description{
\code{build_quarto_site()} is a wrapper for \code{\link[quarto:quarto_render]{quarto::quarto_render()}} that also updates \code{README.md}
and optionally deletes the Quarto \verb{_freeze} directory to fully re-render the site.
}
\details{
When run, \code{build_quarto_site()} calls:
\enumerate{
\item \code{\link[devtools:build_rmd]{devtools::build_readme()}}
\item \code{\link[=unfreeze]{unfreeze()}} (if \code{unfreeze = TRUE})
\item \code{\link[quarto:quarto_render]{quarto::quarto_render()}}
}
}
