% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_rdev_site}
\alias{build_rdev_site}
\title{Build rdev Site}
\usage{
build_rdev_site(pkg = ".", ...)
}
\arguments{
\item{pkg}{Path to package. Currently, only \code{pkg = "."} is supported.}

\item{...}{additional arguments passed to \code{\link[pkgdown:build_site_github_pages]{pkgdown::build_site_github_pages()}} (not implemented)}
}
\description{
\code{build_rdev_site()} is a wrapper for \code{\link[pkgdown:build_site_github_pages]{pkgdown::build_site_github_pages()}} optimized for rdev
workflow that updates \code{README.md} and performs a clean GitHub pages build using \code{pkgdown}.
}
\details{
When run, \code{build_rdev_site()} calls:
\enumerate{
\item \code{\link[devtools:build_rmd]{devtools::build_readme()}}
\item \code{\link[pkgdown:build_site_github_pages]{pkgdown::build_site_github_pages()}} with \code{install = TRUE} and \code{new_process = TRUE}
}
}
\section{Continuous Integration}{

Both \code{\link[=build_rdev_site]{build_rdev_site()}} and \code{\link[=build_analysis_site]{build_analysis_site()}} are meant to be used as part of a CI/CD
workflow, and temporarily set the environment variable \code{CI == "TRUE"} so that the build will
fail when non-internal topics are not included on the reference index page per
\code{\link[pkgdown:build_reference]{pkgdown::build_reference()}}.
}

