% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{cfile}
\alias{cfile}
\alias{cfile,lockedmod-method}
\alias{cfile,mrgmod-method}
\title{Return the name of the model specification file.}
\usage{
cfile(x, ...)

\S4method{cfile}{mrgmod}(x, ...)

\S4method{cfile}{lockedmod}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}
}
\description{
Return the name of the model specification file.
}

