% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{check_renv}
\alias{check_renv}
\title{Check renv}
\usage{
check_renv(update = rlang::is_interactive(), cooldown = 7)
}
\arguments{
\item{update}{run \code{\link[renv:update]{renv::update()}}}

\item{cooldown}{Number of days ago to set the PPM snapshot date.}
}
\description{
Runs \code{\link[renv:renv-package]{renv}} \code{\link[renv:status]{status()}}, \code{\link[renv:clean]{clean()}},
\code{\link[renv:vulns]{vulns()}} (using \code{\link[=renv_vulns]{renv_vulns()}}), and optionally
\code{\link[renv:update]{update()}}
}
\examples{
\dontrun{
check_renv()
check_renv(update = FALSE)
}
}
