% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\title{Local CI}
\usage{
ci(
  renv = TRUE,
  missing = TRUE,
  pkgdown = TRUE,
  styler = NULL,
  lintr = TRUE,
  document = TRUE,
  normalize = TRUE,
  extra = TRUE,
  spelling = TRUE,
  urls = TRUE,
  url_check_fail = getOption("rdev.url_check.fail", default = TRUE),
  rcmdcheck = TRUE
)
}
\arguments{
\item{renv}{check \code{\link[renv:status]{renv::status()}} and report on \code{\link[renv:vulns]{renv::vulns()}} (as YAML)}

\item{missing}{run \code{\link[=missing_deps]{missing_deps()}}}

\item{pkgdown}{check \code{\link[pkgdown:check_pkgdown]{pkgdown::check_pkgdown()}} if \verb{_pkgdown.yml} exists}

\item{styler}{style all files using \code{\link[=style_all]{style_all()}}, see details}

\item{lintr}{lint all files using \code{\link[=lint_all]{lint_all()}}}

\item{document}{run \code{\link[devtools:document]{devtools::document()}}}

\item{normalize}{run \code{\link[desc:desc_normalize]{desc::desc_normalize()}}}

\item{extra}{run \code{\link[=extra_deps]{extra_deps()}}}

\item{spelling}{update spelling \code{\link[spelling:wordlist]{WORDLIST}}}

\item{urls}{validate URLs with \code{\link[=url_check]{url_check()}} and \code{\link[=html_url_check]{html_url_check()}}}

\item{url_check_fail}{throw an error and stop \code{ci()} when one or more URLs are flagged by
\code{\link[=url_check]{url_check()}}}

\item{rcmdcheck}{run \verb{R CMD check} using:
\code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck(args = "--no-manual", error_on = "warning")}}}
}
\description{
Run continuous integration tests locally.
}
\details{
If \code{\link[renv:status]{renv::status()}} is not synchronized, \code{ci()} will stop.

If \code{\link[=missing_deps]{missing_deps()}} returns any missing dependencies, \code{ci()} will stop.

\code{\link[pkgdown:check_pkgdown]{pkgdown::check_pkgdown()}} will halt \code{ci()} with an error if \verb{_pkgdown.yml} is invalid.

If \code{styler} is set to \code{NULL} (the default), \code{\link[=style_all]{style_all()}} will be run only if there are no
uncommitted changes to git. Setting the value to \code{TRUE} or \code{FALSE} overrides this check.

If \code{\link[=lint_all]{lint_all()}} finds any lints, \code{ci()} will stop and open the RStudio markers pane.

Output from \code{missing}, \code{extra}, and \code{urls} is printed as a \link[tibble:tibble]{tibble} for
improved readability in the console.
}
\examples{
\dontrun{
ci()
ci(styler = TRUE)
ci(styler = FALSE, rcmdcheck = FALSE)
}
}
