% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complog.R
\name{cleanso}
\alias{cleanso}
\title{Clean up model shared objects.}
\usage{
cleanso(x, where = soloc(x))
}
\arguments{
\item{x}{model object}
}
\description{
Clean up model shared objects.
}
\details{
\code{cleanso} removes (deletes) shared objects from the model compile directory and 
attempts to unload shared objects which appear to be loaded.
}

