% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{cmt}
\alias{cmt}
\alias{cmt,mrgmod-method}
\title{Get the names of model compartments.}
\usage{
cmt(x, ...)

\S4method{cmt}{mrgmod}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}
}
\description{
Get the names of model compartments.
}
\examples{
mod <- mrgsolve:::house()

cmt(mod)


}

