% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complog.R
\name{complog}
\alias{comp_forget}
\alias{complog}
\title{Functions for viewing and manipulating the compilation log.}
\usage{
complog(full = FALSE)

comp_forget(x)
}
\arguments{
\item{full}{show a full display}

\item{x}{not used}
}
\description{
\code{complog} displays all of the models in the compilation log.  \code{comp_forget} removes models from the compilation log and attempts to unload the corresponding shared object.
}

