% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rdev.R
\name{create_github_repo}
\alias{create_github_repo}
\title{Create rdev GitHub repository}
\usage{
create_github_repo(repo_name, repo_desc = "", host = NULL)
}
\arguments{
\item{repo_name}{The name of the GitHub repository to create}

\item{repo_desc}{The description of the GitHub repository to create}

\item{host}{GitHub host to target, passed to the \code{.api_url} argument of
\code{\link[gh:gh]{gh::gh()}}. If unspecified, gh defaults to "https://api.github.com",
although gh's default can be customised by setting the GITHUB_API_URL
environment variable.

For a hypothetical GitHub Enterprise instance, either
"https://github.acme.com/api/v3" or "https://github.acme.com" is
acceptable.}
}
\value{
return value from \code{\link[gh:gh]{gh::gh()}} creating the repository, invisibly
}
\description{
Create new GitHub repository following rdev conventions in the active user's account using
\code{\link[gh:gh]{gh::gh()}}, activates Dependabot alerts, Dependabot security updates, clones the repository
locally with \code{\link[usethis:create_from_github]{usethis::create_from_github()}}, and creates a basic package using
\code{\link[usethis:create_package]{usethis::create_package()}}. If running interactively on macOS, the repository will
automatically be opened in RStudio, GitHub Desktop, and the default browser.
}
