% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\docType{methods}
\name{data_set}
\alias{data_set}
\alias{data_set,mrgmod,ANY-method}
\alias{data_set,mrgmod,data.frame-method}
\title{Set the \code{data} argument for \code{mrgsim}.}
\usage{
data_set(x, data, ...)

\S4method{data_set}{mrgmod,data.frame}(x, data, subset = TRUE,
  select = TRUE, rename = NULL, ...)

\S4method{data_set}{mrgmod,ANY}(x, data, ...)
}
\arguments{
\item{x}{model object}

\item{data}{data set}

\item{...}{passed along}

\item{subset}{passed to \code{dplyr::filter_}; retain only certain rows in the data set}

\item{select}{passed to \code{dplyr::select_}; retain only certain columns in the data set}

\item{rename}{parsed and passed to \code{dplyr::rename_}; change the names of columns in the data set}
}
\description{
Set the \code{data} argument for \code{mrgsim}.
}
\details{
Input data sets are \code{R} data frames that can include columns with any valid name, however columns with selected names are recognized by \code{mrgsolve} and incorporated into the simulation.

\code{ID} specifies the subject ID and is required for every input data set.

When columns have the same name as parameters (\code{$PARAM} in the model specification file), the values in those columns will be used to update the corresponding parameter as the simulation progresses.

Input data set may include the following columns related to PK dosing events: \code{time}, \code{cmt}, \code{amt}, \code{rate},
\code{ii}, \code{addl}, \code{ss}.  \code{time} and \code{cmt} (and \code{ID}) are required columns in the input data set.  \code{time} is the observation or event time, \code{cmt} is the compartment number (see \code{\link{init}}), \code{amt} is the dosing amount, \code{rate} is the infusion rate, \code{ii} is the dosing interval, \code{addl} specifies additional doses to administer, and \code{ss} is a flag for steady state dosing.  These column names operate similarly to other non-linear mixed effects modeling software, but note that (except for \code{ID}) the column names related to PK dosing must be lower case.

Only numeric data can be brought in to the problem.  Any non-numeric data columns will be dropped with warning.

See \code{\link{exdatasets}} for different example data sets.
}
\examples{

data <- expand.ev(ID=1:3, amt=c(10,20))

data <- expand.ev(amt=c(10,20), rate=c(1,2))


}

