% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missing_deps}
\alias{missing_deps}
\alias{extra_deps}
\title{Check dependencies}
\usage{
missing_deps()

extra_deps()
}
\value{
data.frame from either \code{\link[renv:dependencies]{renv::dependencies()}} or \code{\link[desc:desc_get_deps]{desc::desc_get_deps()}}.
}
\description{
Check dependencies in DESCRIPTION.

\code{missing_deps()} reports \code{\link[renv:dependencies]{renv::dependencies()}} not in DESCRIPTION.

\code{extra_deps()} reports \code{\link[desc:desc_get_deps]{desc::desc_get_deps()}} not found by renv.

The current package (\code{\link[pkgload:packages]{pkgload::pkg_name(".")}}) is automatically removed
from \code{\link[renv:dependencies]{renv::dependencies()}}.
}
