% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{design}
\alias{design}
\title{Set observation designs for the simulation.}
\usage{
design(x, descol = character(0), ..., deslist = list())
}
\arguments{
\item{x}{model object}

\item{descol}{the \code{idata} column name for design assignment}

\item{...}{\code{tgrid} or \code{tgrids} objects or \code{numeric} vector}

\item{deslist}{a list of \code{tgrid} or \code{tgrids} objects or \code{numeric} vector to be used in place of ...}
}
\description{
Set observation designs for the simulation.
}

