% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{dllname}
\alias{dllname}
\alias{dllname,mrgmod-method}
\alias{package}
\alias{package,mrgmod-method}
\title{Return the model name.}
\usage{
dllname(x, ...)

\S4method{dllname}{mrgmod}(x, ...)

package(x, ...)

\S4method{package}{mrgmod}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}
}
\description{
Return the model name.
}
\examples{
mod <- mrgsolve:::house()
dllname(mod)

}
\seealso{
\code{\link{model}}
}

