% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_license}
\alias{get_license}
\title{Get license option}
\usage{
get_license()
}
\value{
license string, one of \code{c("mit", "gpl", "lgpl", "proprietary")}
}
\description{
Retrieve and validate \code{rdev.license} option.
}
\details{
\code{rdev.license} must be one of \code{c("mit", "gpl", "lgpl", "proprietary")}, and defaults to \code{"mit"}.
If \code{rdev.license} is \code{"proprietary"}, \code{rdev.license.copyright} (the name of the copyright
holder) must also be set.
}
