% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{get_release}
\alias{get_release}
\title{Get release details}
\usage{
get_release(pkg = ".", filename = "NEWS.md")
}
\arguments{
\item{pkg}{Path to package. Currently, only \code{pkg = "."} is supported.}

\item{filename}{name of file containing release notes, defaults to \code{NEWS.md}.}
}
\value{
list containing the version and release notes from the first release contained in
\code{NEWS.md}
}
\description{
Extract release version and release notes from \code{NEWS.md}.
}
\details{
\code{get_release()} assumes that \code{NEWS.md} contains markdown release notes, with each release header
of the format: \code{"# <package> <version>"} followed by the release notes, and expects the first
line of \code{NEWS.md} to be a release header.
}
