% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{CMT}
\alias{CMT}
\title{Parse $CMT block.}
\usage{
CMT(x, env, annotated = FALSE, pos = 1, object = NULL, ...)
}
\arguments{
\item{x}{data}

\item{env}{parse environment}

\item{annotated}{logical}

\item{pos}{block position}

\item{...}{passed}
}
\description{
Parse $CMT block.
}

