% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{FIXED}
\alias{FIXED}
\title{Parse \code{$FIXED} block.}
\usage{
FIXED(x, env, annotated = FALSE, pos = 1, ...)
}
\arguments{
\item{x}{data}

\item{env}{parse environment}

\item{annotated}{logical}

\item{pos}{parse position}

\item{...}{passed}
}
\description{
Parse \code{$FIXED} block.
}

