% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{THETA}
\alias{THETA}
\title{Parse \code{$THETA} block.}
\usage{
THETA(x, env, annotated = FALSE, pos = 1, object = NULL, name = "THETA",
  ...)
}
\arguments{
\item{x}{data}

\item{env}{parse environment}

\item{annotated}{logical}

\item{pos}{parse position}

\item{object}{the (character) name of a \code{list} in \code{$ENV} to use for block output}

\item{name}{character prefix for parameter names}

\item{...}{passed}
}
\description{
Parse \code{$THETA} block.
}
\details{
If \code{object} is specified, it must be the name of a named list of
\code{THETA} values.
}

