% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{idata}
\alias{idata}
\title{Create an idata data set}
\usage{
idata(..., KEEP.OUT.ATTRS = FALSE, stringsAsFactors = FALSE)
}
\arguments{
\item{...}{passed to expand.grid}

\item{KEEP.OUT.ATTRS}{passed to expand.grid}

\item{stringsAsFactors}{passed to expand.grid}
}
\description{
Create an idata data set
}

