% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\docType{methods}
\name{idata_set}
\alias{idata_set}
\alias{idata_set,mrgmod,ANY-method}
\alias{idata_set,mrgmod,data.frame-method}
\title{Set the \code{idata} argument for \code{mrgsim}.}
\usage{
idata_set(x, data, ...)

\S4method{idata_set}{mrgmod,data.frame}(x, data, subset = TRUE,
  select = TRUE, ...)

\S4method{idata_set}{mrgmod,ANY}(x, data, ...)
}
\arguments{
\item{x}{model object}

\item{data}{a data set coercable to data.frame}

\item{...}{passed along}

\item{subset}{passed to \code{dplyr::filter_}}

\item{select}{passed to \code{dplyr::select_}}
}
\description{
Set the \code{idata} argument for \code{mrgsim}.
}

