% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{initials}
\alias{initials}
\alias{initials,parset-method}
\title{Generate initial estimates from a parset.}
\usage{
initials(x, ...)

\S4method{initials}{parset}(x, ...)
}
\arguments{
\item{x}{a parset object}

\item{...}{not used}
}
\description{
Generate initial estimates from a parset.
}

