% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsims.R
\docType{methods}
\name{label}
\alias{label}
\alias{label,mrgsims-method}
\title{Label simulation output.}
\usage{
label(x, ...)

\S4method{label}{mrgsims}(x, ...)
}
\arguments{
\item{x}{mrgsims object}

\item{...}{name=value pairs; value must be numeric.}
}
\description{
Attaches a named column to the simulation output with a single numeric value.
}

