% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsims.R
\docType{methods}
\name{limit}
\alias{limit}
\alias{limit,mrgsims-method}
\title{Limit the scope of simulated output.}
\usage{
limit(x, ...)

\S4method{limit}{mrgsims}(x, subset, select = TRUE, ...)
}
\arguments{
\item{x}{mrgsims or batch mrgsims object}

\item{...}{passed along}

\item{subset}{rows to keep}

\item{select}{columns to keep}
}
\description{
Limit the scope of simulated output.
}

