% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{lint_all}
\alias{lint_all}
\title{Lint all files}
\usage{
lint_all(
  pattern = "(?i)[.](r|rmd|qmd|rnw|rhtml|rpres|rrst|rtex|rtxt)$",
  exclusions = list("renv", "packrat", "R/RcppExports.R"),
  ...
)
}
\arguments{
\item{pattern}{regex pattern for files, by default it will take files with any of the extensions
.R, .Rmd, .qmd, .Rnw, .Rhtml, .Rpres, .Rrst, .Rtex, .Rtxt, ignoring case.}

\item{exclusions}{exclusions for \code{\link[lintr:exclude]{exclude()}}, relative to the package path.}

\item{...}{
  Arguments passed on to \code{\link[lintr:lint_dir]{lintr::lint_dir}}
  \describe{
    \item{\code{filename}}{Either the filename for a file to lint, or a character string of inline R code for linting.
The latter (inline data) applies whenever \code{filename} has a newline character (\\n).}
    \item{\code{linters}}{A named list of linter functions to apply. See \link[lintr]{linters} for a full list of default and available
linters.}
    \item{\code{cache}}{When logical, toggle caching of lint results. If passed a character string, store the cache in this
directory.}
    \item{\code{parse_settings}}{Logical. Whether to try and parse the \link[lintr:read_settings]{settings}. Otherwise, the
\code{\link[lintr:default_settings]{default_settings()}} are used. \code{TRUE} by default when linting files, as opposed to \verb{text=}.}
    \item{\code{text}}{Optional argument for supplying a string or lines directly, e.g. if the file is already in memory or
linting is being done ad hoc.}
    \item{\code{path}}{For the base directory of the project (for \code{lint_dir()}) or
package (for \code{lint_package()}).}
    \item{\code{relative_path}}{if \code{TRUE}, file paths are printed using their path relative to the base directory.
If \code{FALSE}, use the full absolute path.}
    \item{\code{show_progress}}{Logical controlling whether to show linting progress with
\code{\link[cli:cli_progress_along]{cli::cli_progress_along()}}. The default behavior is to show progress in \code{\link[=interactive]{interactive()}} sessions
not running a testthat suite.}
  }}
}
\value{
A list of lint objects.
}
\description{
Lint all files in a project. Implemented as a wrapper for \code{\link[lintr:lint_dir]{lintr::lint_dir()}}.
}
\examples{
\dontrun{
lint_all()
lint_all("analysis")
}
}
