% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{local_temppkg}
\alias{local_temppkg}
\title{Temporary package}
\usage{
local_temppkg(dir = fs::file_temp(), type = "usethis", env = parent.frame())
}
\arguments{
\item{dir}{Path to package directory, created if necessary, defaults to \code{\link[fs:file_temp]{fs::file_temp()}}.}

\item{type}{type of package to create, one of "usethis" - \code{\link[usethis:create_package]{usethis::create_package()}},
"rdev" - \code{\link[=use_rdev_package]{use_rdev_package()}}, or "analysis" - \code{\link[=use_analysis_package]{use_analysis_package()}}.}

\item{env}{Environment passed to \code{\link[withr:defer]{withr::defer()}}, defaults to \code{\link[=parent.frame]{parent.frame()}}.}
}
\value{
Path to temporary package directory.
}
\description{
Temporarily create a rdev or R analysis package, which is automatically removed afterwards.
}
\details{
Used internally for testing rdev automation. Based on the
\href{https://testthat.r-lib.org/articles/test-fixtures.html#case-study-usethis}{usethis case study}
from \code{\link[testthat:testthat-package]{testthat}}.
}
\examples{
\dontrun{
test_that("local_temppkg creates a directory", {
  dir <- usethis::ui_silence(local_temppkg())
  expect_true(fs::dir_exists(dir))
})
}
}
\seealso{
\link[withr:withr]{withr::withr-package}
}
