% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{lower2matrix}
\alias{lower2matrix}
\title{Create a square numeric matrix from the lower-triangular elements.}
\usage{
lower2matrix(x, prefix = NULL)
}
\arguments{
\item{x}{numeric data}

\item{prefix}{used to generate column names}
}
\value{
a square symmetric numeric matrix with column names
}
\description{
Create a square numeric matrix from the lower-triangular elements.
}

