% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_matlist.R, R/class_derived.R, R/class_modlist.R
\docType{class}
\name{matlist-class}
\alias{matlist-class}
\alias{modlist-class}
\alias{omegalist-class}
\alias{sigmalist-class}
\title{S4 class matlist.}
\description{
S4 class matlist.

S4 class matlist.
}

