% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{mcode}
\alias{mcode}
\title{TO BE REMOVED 4/29/16
Write, compile, and load model code.}
\usage{
mcode(model, code, project = tempdir(), ...)
}
\arguments{
\item{model}{model name}

\item{code}{character string specifying a \code{mrgsolve} model}

\item{project}{project name}

\item{...}{passed to \code{\link{mread}}}
}
\description{
This is a convenience function that ultimately calls \code{\link{mread}}.
}
\details{
Note that the arguments are in slightly different order than \code{\link{mread}}.  The default \code{project} is \code{tempdir()}.
}
\examples{

code <- '
$CMT DEPOT CENT
$ADVAN2
$MAIN
pred_CL = 1;
pred_VC= 20;
'

mod <- mcode("example",code)

}

