% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{merge_release}
\alias{merge_release}
\title{Merge staged GitHub release}
\usage{
merge_release(pkg = ".", filename = "NEWS.md", host = getOption("rdev.host"))
}
\arguments{
\item{pkg}{path to package. Currently, only \code{pkg = "."} is supported.}

\item{filename}{name of file containing release notes, defaults to \code{NEWS.md}.}

\item{host}{GitHub host to target, passed to the \code{.api_url} argument of
\code{\link[gh:gh]{gh::gh()}}. If unspecified, gh defaults to "https://api.github.com",
although gh's default can be customised by setting the GITHUB_API_URL
environment variable.

For a hypothetical GitHub Enterprise instance, either
"https://github.acme.com/api/v3" or "https://github.acme.com" is
acceptable.}
}
\value{
list containing results of pull request merge and GitHub release, invisibly
}
\description{
Merge a pull request staged with \code{\link[=stage_release]{stage_release()}} and create a new release on GitHub.
}
\details{
Manually verify that all status checks have completed before running, as \code{merge_release()}
doesn't currently validate that status checks are successful.

When run, \code{merge_release()}:
\enumerate{
\item Checks for uncommitted changes and stops if any exist using \code{\link[gert:git_status]{gert::git_status()}}
\item Determines the staged release title from \code{NEWS.md} using \code{\link[=get_release]{get_release()}}
\item Selects the GitHub pull request that matches the staged release title, stops if there is more
or less than one matching PR using \code{\link[gh:gh]{gh::gh()}}
\item Verifies the staged pull request is ready to be merged by checking the locked, draft,
mergeable, and rebaseable flags
\item Merges the pull request into the default branch using "Rebase and merge" using \code{\link[gh:gh]{gh::gh()}}
\item Deletes the pull request branch remotely and locally using \code{\link[gh:gh]{gh::gh()}} and
\code{\link[gert:git_branch_delete]{gert::git_branch_delete()}}
\item Updates the default branch with \code{\link[gert:git_pull]{gert::git_pull()}}
\item Adds the version tag to the \code{DESCRIPTION} commit with the message \code{"GitHub release <version>"}
with \code{\link[gert:git_tag_create]{gert::git_tag_create()}} and pushes using \code{\link[gert:git_tag_push]{gert::git_tag_push()}}
\item Create the GitHub release from the newly created tag, with the name \code{"<version>"} and the
release notes in the body, using \code{\link[gh:gh]{gh::gh()}}
}
}
\section{Host}{

Set the \code{rdev.host} option when using a GitHub Enterprise server:
\code{options(rdev.host = "https://github.example.com/api/v3")}
}

