% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintern.R
\name{mintern}
\alias{mintern}
\title{Load a library model.}
\usage{
mintern(model, tryload = TRUE, ...)
}
\arguments{
\item{model}{model name}

\item{tryload}{try to load the model?}

\item{...}{passed to update}
}
\description{
Load a library model.
}
\examples{
mod <- mread("pk1cmt",.modl())
mod <- mread("pk2cmt",.modl()) 
mod <- mread("pk3cmt",.modl()) 
mod <- mread("irm1",.modl()) 
mod <- mread("irm2",.modl()) 
mod <- mread("irm3",.modl()) 
mod <- mread("irm4",.modl())
mod <- mread("emax",.modl())
mod <- mread("effect",.modl())
mod <- mread("tmdd",.modl()) 

}

