% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{model}
\alias{model}
\alias{model,mrgmod-method}
\title{Return the model name.}
\usage{
model(x, ...)

\S4method{model}{mrgmod}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}
}
\description{
Return the model name.
}
\examples{
mod <- mrgsolve:::house()
model(mod)
}

