% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{modelparse}
\alias{modelparse}
\title{Parse model specification text.}
\usage{
modelparse(txt, split = FALSE, ...)
}
\arguments{
\item{txt}{model specification text}

\item{split}{logical}

\item{...}{arguments passed along}
}
\description{
Parse model specification text.
}

