% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_modlist.R
\docType{methods}
\name{$,modlist-method}
\alias{$,modlist-method}
\alias{modlist}
\title{Create a modlist object.}
\usage{
\S4method{$}{modlist}(x, name)

modlist(project = ".", soloc = tempdir(), prefix = "")
}
\arguments{
\item{name}{model to take; used with \code{$}}
}
\description{
Create a modlist object.
}

