% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolveModels.R
\docType{package}
\name{modmrg}
\alias{modmrg}
\alias{modmrg-package}
\title{Precompiled models for use with \code{mrgsolve}.}
\description{
Precompiled models for use with \code{mrgsolve}.
}
\section{Package help}{

\itemize{
 \item  Package \href{00Index.html}{index}, including a listing of all functions
}
}

\section{Types of models}{


Currently available are:
\itemize{
\item Traditional \code{\link{pk}} models
\item \code{\link{pkpd}} models
\item \code{\link{tmdd}} models
\item \code{\link{viral}} dynamics models
}
}
\examples{
mod <- pk1cmt()
mod <- pk2cmt()
mod <- pk3cmt()
mod <- irm1()
mod <- irm2()
mod <- irm3()
mod <- irm4()
mod <- pd_effect()
mod <- emax()
mod <- tmdd()
mod <- viral1()
mod <- viral2()

}

