% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{mread}
\alias{mread}
\title{Read a model specification file}
\usage{
mread(model = character(0), project = getwd(), code = NULL, udll = TRUE,
  ignore.stdout = TRUE, raw = FALSE, compile = TRUE, audit = FALSE,
  quiet = getOption("mrgsolve_mread_quiet", FALSE), check.bounds = FALSE,
  warn = TRUE, soloc = tempdir(), preclean = FALSE, ...)
}
\arguments{
\item{model}{model name}

\item{project}{working directory}

\item{code}{a character string with model specification code}

\item{udll}{use unique name for shared object}

\item{ignore.stdout}{passed to system call for compiling model}

\item{raw}{if TRUE, return a list of raw output}

\item{compile}{try to compile the model and load the shared object}

\item{audit}{check the model specification file for errors}

\item{quiet}{don't print messages when compiling}

\item{check.bounds}{check boundaries of parameter list}

\item{warn}{logical; if \code{TRUE}, print warning messages that may arise}

\item{soloc}{directory where model shared object is stored}

\item{preclean}{logical; if \code{TRUE}, compilation artifacts are cleaned up first}

\item{...}{passed along}
}
\description{
Read a model specification file
}
\examples{


code <- '
$PARAM CL = 1, VC = 5
$CMT CENT
$ODE dxdt_CENT = -(CL/VC)*CENT;
'

mod <- mcode("ex_mread",code)

mod

mod \%>\% init(CENT=1000) \%>\% mrgsim \%>\% plot

}

