% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve.R
\name{mrgindata}
\alias{mrgindata}
\alias{mrgindata.data.frame}
\alias{mrgindata.matrix}
\title{Prepare input data.frame or matrix}
\usage{
mrgindata(x, ...)

\method{mrgindata}{data.frame}(x, m = NULL, verbose = FALSE,
  quiet = FALSE, ...)

\method{mrgindata}{matrix}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{data.frame or matrix}

\item{...}{additional arguments}

\item{m}{object that inherits from mrgmod}

\item{verbose}{logical}

\item{quiet}{if \code{TRUE}, messages will be suppressed}
}
\value{
a matrix with non-numeric columns dropped; if x is a data.frame with character \code{cmt} column comprised of valid compartment names and \code{m} is a model object,
the \code{cmt} column will be converted to the corresponding compartment number.
}
\description{
Prepare input data.frame or matrix
}

