% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{mrgsolve_example}
\alias{mrgsolve_example}
\title{Extract example model from system library}
\usage{
mrgsolve_example(model = c("pkExample", "pkpdExample", "firstmodeExample",
  "viralExample", "popExample"), project = getwd(), overwrite = FALSE,
  quiet = FALSE, ...)
}
\arguments{
\item{model}{name of model}

\item{project}{working directory}

\item{overwrite}{passed to file.copy}

\item{quiet}{don't print any status messages to the screen}

\item{...}{additional arguments}
}
\description{
Extract example model from system library
}
\examples{
## example("mrgsolve_example", package="mrgsolve")

mrgsolve_example("pkpdExample", project=getwd())

mod <- mread("pkpdExample", project=getwd()) \%>\% ev(amt=1000, time=0, cmt=1)

see(mod)

out <- mod \%>\% mrgsim(end=48,delta=0.1)

out

plot(out)

out <- mod \%>\%
  ev(amt=1000, ii=24, cmt=1, addl=10)  \%>\%
  mrgsim(end=300)

plot(out)
plot(out, CP~time)

}

