% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{neq}
\alias{neq}
\alias{neq,mrgmod-method}
\title{Return the number of compartments / equations.}
\usage{
neq(x, ...)

\S4method{neq}{mrgmod}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}
}
\description{
Return the number of compartments / equations.
}
\examples{
mod <- mrgsolve:::house()
neq(mod)

}

