% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paths.R
\name{normalize_path}
\alias{normalize_path}
\title{Normalize Path}
\usage{
normalize_path(path, winslash = "/", ...)
}
\arguments{
\item{path}{character vector of file paths.}

\item{winslash}{the separator to be used on Windows -- ignored
    elsewhere.  Must be one of \code{c("/", "\\\\")}.}

\item{...}{
  Arguments passed on to \code{\link[base:normalizePath]{base::normalizePath}}
  \describe{
    \item{\code{mustWork}}{logical: if \code{TRUE} then an error is given if the result
    cannot be determined; if \code{NA} then a warning.}
  }}
}
\value{
A character vector.

  If an input is not a real path the result is system-dependent (unless
  \code{mustWork = TRUE}, when this should be an error).  It will be
  either the corresponding input element or a transformation of it into
  an absolute path.

  Converting to an absolute file path can fail for a large number of
  reasons.  The most common are
  \itemize{
    \item One of more components of the file path does not exist.
    \item A component before the last is not a directory, or there is
    insufficient permission to read the directory.
    \item For a relative path, the current directory cannot be
    determined.
    \item A symbolic link points to a non-existent place or links form a
    loop.
    \item The canonicalized path would be exceed the maximum supported
    length of a file path.
  }
}
\description{
Simple wrapper around \code{\link[base:normalizePath]{base::normalizePath()}} with defaults that set \code{winslash}
to \code{"/"} for better cross-platform interoperability.
}
