% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{numeric2diag}
\alias{numeric2diag}
\title{Create a diagonal numeric matrix from diagonal elements.}
\usage{
numeric2diag(x, prefix = NULL)
}
\arguments{
\item{x}{numeric data}

\item{prefix}{used to generate column names}
}
\value{
a numeric diagonal matrix
}
\description{
Create a diagonal numeric matrix from diagonal elements.
}

