% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{obsaug}
\alias{obsaug}
\title{Set the \code{obsaug} argument for \code{mrgsim}.}
\usage{
obsaug(x, value = TRUE, ...)
}
\arguments{
\item{x}{model object}

\item{value}{the value for \code{obsaug}}

\item{...}{passed along}
}
\description{
Set the \code{obsaug} argument for \code{mrgsim}.
}

