% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_files}
\alias{open_files}
\title{Open rdev files}
\usage{
open_files(files = c("TODO.md", "NEWS.md", "README.Rmd", "DESCRIPTION"))
}
\arguments{
\item{files}{vector of files to open.}
}
\value{
named character vector of files opened.
}
\description{
Open a standard set of files for editing in RStudio.
}
\details{
By default, \code{open_files()} opens four documents in RStudio: \code{TODO.md}, \code{NEWS.md}, \code{README.Rmd},
and \code{DESCRIPTION}.

\code{open_files()} will stop with an error if RStudio is not running.
}
