% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_downloads}
\alias{package_downloads}
\title{Summarize package downloads}
\usage{
package_downloads(packages, when = "last-month")
}
\arguments{
\item{packages}{A character vector of the packages to query.}

\item{when}{The period to summarize, one of \code{last-day}, \code{last-week} or \code{last-month}
(the default).}
}
\value{
A data frame containing the total number of downloads by package for the specified
period, sorted by popularity.
}
\description{
A wrapper for \code{\link[cranlogs:cran_downloads]{cranlogs::cran_downloads()}} that summarizes the number of package downloads from
the RStudio CRAN mirror.
}
\details{
By default, the summary is for the last month.
}
