% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{pars}
\alias{pars}
\alias{pars,mrgmod-method}
\title{Return the names of model parameters.}
\usage{
pars(x, ...)

\S4method{pars}{mrgmod}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}
}
\description{
Return the names of model parameters.
}
\examples{
mod <- mrgsolve:::house()
pars(mod)

}

