% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolveModels.R
\name{pd_effect}
\alias{pd_effect}
\title{Effect compartment model.}
\usage{
pd_effect(...)
}
\arguments{
\item{...}{passed to \code{\link{update}}}
}
\description{
Effect compartment model.
}
\details{
See \code{\link{pkpd_details}} for more detailed 
descriptions of parameters and compartments.
}

