% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolveModels.R
\name{pk}
\alias{pk}
\alias{pk1cmt}
\alias{pk1cmt_pop}
\alias{pk2cmt}
\alias{pk3cmt}
\title{Pharmacokinetic models.}
\usage{
pk1cmt(...)

pk2cmt(...)

pk3cmt(...)

pk1cmt_pop(...)
}
\arguments{
\item{...}{passed to update}
}
\value{
an object of class \code{packmod}
}
\description{
Pharmacokinetic models.
}
\details{
See \code{\link{pkpd_details}} for more detailed descriptions of parameters and compartments.

The \code{pk1cmt} model is parameterized in terms of \code{CL}, \code{VC}, \code{KA1} and \code{KA2} and uses compartments \code{EV1},
\code{EV2}, and \code{CENT}.  The \code{pk2cmt} model adds a \code{PERIPH} compartment and parameters \code{Q} and \code{VP} to that of the
one-compartment model.  Likewise, the three-compartment model (\code{pk3cmt}) adds \code{PERIPH2} and parameters \code{Q2} and \code{VP2} to
that of the two-compartment models.  All pk models also have parameters \code{VMAX} (defaulting to zero, no non-linear clearance) and \code{KM}.
}
\section{Model description}{

All pk models have two extravascular dosing compartments and potential for linear and nonlinear clearance.
\itemize{
 \item{\code{pk1cmt}}: one compartment pk model
 \item{\code{pk2cmt}}: two compartment pk  model
 \item{\code{pk3cmt}}: three compartment pk model
}
}
\examples{

## example("pk", package="modmrg")

mod <- pk1cmt(end=180)

events <- ev(amt=1000, cmt=1, addl=4, ii=24)


mod
see(mod)
param(mod)
init(mod)
events

out <- mod \%>\% ev(events) \%>\% mrgsim(delta=0.25)

plot(out, CP~.)


mod <- pk1cmt_pop(end=180,delta=0.25)

omega  <-  omat(dmat(0.1,0.3,0.2, 0.6))

out <- mod \%>\% ev(events) \%>\%  mrgsim(omega=omega, idata=data.frame(ID=1:10))

plot(out, CP~.)

}

