% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{PKMODEL}
\alias{PKMODEL}
\title{Parse data from \code{$PKMODEL}}
\usage{
PKMODEL(ncmt = 1, depot = FALSE, trans = pick_trans(ncmt, depot), ...)
}
\arguments{
\item{ncmt}{number of compartments}

\item{depot}{logical indicating whether to add depot compartment}

\item{trans}{the parameterization for the PK model}

\item{...}{not used}
}
\description{
Parse data from \code{$PKMODEL}
}

