% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knobs.R
\docType{methods}
\name{plot,batch_mrgsims,missing-method}
\alias{plot,batch_mrgsims,formula-method}
\alias{plot,batch_mrgsims,missing-method}
\title{Plot method for mrgsims objects.}
\usage{
\S4method{plot}{batch_mrgsims,missing}(x, yval = variables(x), limit = 9,
  ...)

\S4method{plot}{batch_mrgsims,formula}(x, y, show.grid = TRUE, lwd = 2,
  type = "l", as = "raw", auto.key = list(columns = 1), scales = list(y
  = list(relation = "free")), ...)
}
\arguments{
\item{x}{mrsims object}

\item{yval}{variables to plot}

\item{limit}{maximum number of yval to plot}

\item{...}{arguments passed to xyplot}

\item{y}{a formula passed to xyplot}

\item{show.grid}{print grid in the plot}

\item{lwd}{passed to xyplot}

\item{type}{passed to xyplot}

\item{as}{transformation for every yval that is plotted}

\item{auto.key}{passed to xyplot}

\item{scales}{passed to xyplot}
}
\description{
Plot method for mrgsims objects.
}

