% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsims.R
\docType{methods}
\name{plot_mrgsims}
\alias{plot,mrgsims,formula-method}
\alias{plot,mrgsims,missing-method}
\alias{plot_mrgsims}
\title{Generate a quick plot of simulated data.}
\usage{
\S4method{plot}{mrgsims,missing}(x, limit = 16, ...)

\S4method{plot}{mrgsims,formula}(x, y, limit = 16, show.grid = TRUE,
  as = "raw", outer = TRUE, type = "l", lwd = 2, ylab = "raw value",
  groups = ID, scales = list(y = list(relation = "free")), ...)
}
\arguments{
\item{x}{mrgsims object}

\item{limit}{limit the the number of panels to create}

\item{...}{other arguments passed to xyplot}

\item{y}{formula used for plotting}

\item{show.grid}{logical indicating whether or not to draw panel.grid}

\item{as}{transformations for plotting simulated values}

\item{outer}{passed to xyplot}

\item{type}{passed to xyplot}

\item{lwd}{passed to xyplot}

\item{ylab}{passed to xyplot}

\item{groups}{passed to xyplot}

\item{scales}{passed to xyplot}
}
\description{
Generate a quick plot of simulated data.
}
\details{
Values for \code{as} argument: ;  \code{raw}: raw simulated output;
\code{frac}: each observation normalized to baseline value;
\code{cfb}: change (difference) from baseline;
\code{cfblog}: change from baseline of log10-transformed values;
\code{log10y}: log10 transformation; \code{lny}: natural log transformed.
}
\examples{

mod <- mrgsolve:::house(end=48, delta=0.2) \%>\% init(GUT=1000)

out <- mrgsim(mod)

plot(out)

plot(out, subset=time <=24)

plot(out, GUT+CP~.)

plot(out, CP+RESP~time, col="black", scales="same", lty=2)
}

