% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{plus.ev}
\alias{plus.ev}
\title{Add an events object to a mrgmod object}
\usage{
plus.ev(e1, e2)
}
\arguments{
\item{e1}{mrgmod object}

\item{e2}{events object}
}
\description{
Add an events object to a mrgmod object
}

