% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{print_renv_vulns}
\alias{print_renv_vulns}
\title{Print Vulnerable Packages using YAML}
\usage{
print_renv_vulns(packages = NULL, silent = FALSE)
}
\arguments{
\item{packages}{A vector of package specifications, of the form
\verb{<package>==<version>} or \verb{<package>@<version>}.}

\item{silent}{disable output and only return the list of vulnerable packages (invisibly).}
}
\value{
An \R list of vulnerable packages, invisibly.
}
\description{
Inspect packages using \code{\link[renv:vulns]{renv::vulns()}} and pretty-print vulnerable packages using
\code{\link[yaml:as.yaml]{yaml::as.yaml()}}
}
