% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{project}
\alias{project}
\alias{project,mrgmod-method}
\alias{project,mrgsims-method}
\alias{project,packmod-method}
\title{Return the name of the project directory.}
\usage{
project(x, ...)

\S4method{project}{mrgmod}(x, ...)

\S4method{project}{packmod}(x, ...)

\S4method{project}{mrgsims}(x, ...)
}
\arguments{
\item{x}{model object or mrgsims object}

\item{...}{passed along}
}
\description{
Return the name of the project directory.
}
\examples{
mod <- mrgsolve:::house()
project(mod)

}

