% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{qsim}
\alias{qsim}
\title{A quick simulation function.}
\usage{
qsim(x, e, idata, stime)
}
\arguments{
\item{x}{model object}

\item{e}{event object}

\item{idata}{individual data set}

\item{stime}{a \code{\link{tgrid}} object or any object with \code{\link{stime}} method}
}
\description{
A quick simulation function.
}

