% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{quickstart}
\alias{quickstart}
\title{rdev Quick Start}
\description{
Quick start guide to creating a new rdev or analysis package.
}
\details{
To quickly create and configure a new rdev or analysis package, use the following commands:
\enumerate{
\item With no project open, run \code{\link[=create_github_repo]{create_github_repo()}} to initialize the GitHub R repository
\item Without committing to git, run \code{\link[=init]{init()}} in the newly created project
\item Manually update the Title and Description fields in the \code{DESCRIPTION} file without committing
\item Run \code{\link[=setup_analysis]{setup_analysis()}} or \code{\link[=setup_rdev]{setup_rdev()}} to configure the package as an analysis package or
rdev package respectively.
\item Manually update \code{.gitignore}: remove the \verb{docs/} exclusion and add line breaks
}

After this, the package configuration is complete and ready for development.
}
