% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdev-package.R
\docType{package}
\name{rdev-package}
\alias{rdev-package}
\alias{rdev}
\title{\code{rdev} Package}
\description{
The \code{rdev} package is a toolkit providing functions and features meant to streamline
modern R development and enhance the overall developer experience.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/rdev}
  \item \url{http://docs.jimbrig.com/rdev/}
  \item Report bugs at \url{https://github.com/jimbrig/rdev/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

Authors:
\itemize{
  \item Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})
}

}
