% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_config.R
\name{read_renviron}
\alias{read_renviron}
\title{Read \code{.Renviron} File}
\usage{
read_renviron(path = NULL)
}
\arguments{
\item{path}{The directory path where the \code{.Renviron} file is located. If \code{NULL}, defaults to the current working directory.}
}
\value{
Invisibly returns \code{NULL}. The function is called for its side effect of loading environment variables into the R session.
}
\description{
Simple wrapper around \code{\link[base:readRenviron]{base::readRenviron()}} to read a \code{.Renviron} file from a specified path.
If no path is provided, it defaults to the current working directory.
}
