% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{release_stage}
\alias{release_stage}
\title{Stage a GitHub release}
\usage{
release_stage(pkg = ".", filename = "NEWS.md")
}
\arguments{
\item{pkg}{path to package. Currently, only \code{pkg = "."} is supported.}

\item{filename}{name of file containing release notes, defaults to \code{NEWS.md}.}
}
\description{
Open a GitHub pull request for a new release from \code{NEWS.md}. Approve, merge, and create the
release using \code{release_merge()}.
}
\details{
When run, \code{release_stage()}:
\enumerate{
\item Extracts release version and release notes from \code{NEWS.md} using \code{\link[=get_release]{get_release()}}
\item Validates version conforms to rdev conventions (#.#.#) and release notes aren't empty
\item Verifies that version tag doesn't already exist using \code{\link[gert:git_tag]{gert::git_tag_list()}}
\item Checks for uncommitted changes and stops if any exist using \code{\link[gert:git_diff]{gert::git_diff_patch()}}
}
}
