% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Render an \code{mrgsolve} model.}
\usage{
render(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{not used}

\item{file}{file name to use for building; should have \code{.Rmd} extension}

\item{quiet}{not used}

\item{build}{if \code{TRUE}, the document is built with \code{rmarkdown::render}}
}
\value{
If the document is built, the path to the \code{.pdf} file 
is returned.  If the document is not built, path to 
the \code{.Rmd} file is returned.
}
\description{
Render an \code{mrgsolve} model.
}
\examples{
\dontrun{
mod <- mrgsolve:::house()

mrgsolve:::render(mod,file="house.Rmd")
}


}

