% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsims.R
\docType{methods}
\name{request}
\alias{request}
\alias{request,mrgsims-method}
\title{Return the requested compartments from a simulation run.}
\usage{
request(x, ...)

\S4method{request}{mrgsims}(x, ...)
}
\arguments{
\item{x}{mrgsims object}

\item{...}{passed along}
}
\description{
Return the requested compartments from a simulation run.
}
\seealso{
\code{\link{variables}}
}

