% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{revar}
\alias{revar}
\alias{revar,mrgmod-method}
\alias{revar,mrgsims-method}
\title{Get model random effect variances and covariancnes.}
\usage{
revar(x, ...)

\S4method{revar}{mrgmod}(x, ...)

\S4method{revar}{mrgsims}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}
}
\description{
Get model random effect variances and covariancnes.
}

