% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{Rx}
\alias{Rx}
\title{Create a regimen as event object.}
\usage{
Rx(amt, ii, addl, ...)
}
\arguments{
\item{amt}{dose amount}

\item{ii}{dosing interval}

\item{addl}{additional doses}

\item{...}{passed to \code{\link{ev}}}
}
\description{
Create a regimen as event object.
}

