% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{setup_rdev}
\alias{setup_rdev}
\title{Set up rdev package}
\usage{
setup_rdev()
}
\description{
Set up an rdev package for traditional package development after running \code{\link[=init]{init()}} and updating
Title and Description in \code{DESCRIPTION}, by committing changes and running \code{\link[=use_rdev_pkgdown]{use_rdev_pkgdown()}},
\code{\link[=use_spelling]{use_spelling()}}, \code{\link[=use_codecov]{use_codecov()}}, and \code{\link[=ci]{ci()}}.
}
\details{
\code{setup_rdev()} will stop if \code{\link[rlang:is_interactive]{rlang::is_interactive()}} is \code{FALSE}.
}
\seealso{
\link{quickstart}
}
