% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{shlib}
\alias{shlib}
\alias{shlib,mrgmod-method}
\title{Return information about model compilation.}
\usage{
shlib(x, ...)

\S4method{shlib}{mrgmod}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}
}
\description{
Return information about model compilation.
}

