% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{simre}
\alias{simre}
\alias{simre,mrgmod-method}
\alias{simre,mrgsims-method}
\title{Simulate random effects from model.}
\usage{
simre(x, ...)

\S4method{simre}{mrgmod}(x, seed = NULL, neta = 10, neps = 10, ...)

\S4method{simre}{mrgsims}(x, seed = NULL, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}

\item{seed}{passed to \code{\link{set.seed}}}

\item{neta}{number of etas to simulate}

\item{neps}{number of epsilon values to simulate}
}
\description{
Simulate random effects from model.
}

