% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{sodll}
\alias{sodll}
\alias{sodll,lockedmod-method}
\alias{sodll,mrgmod-method}
\alias{sodll,packmod-method}
\title{Return the name of the shared object file.}
\usage{
sodll(x, ...)

\S4method{sodll}{mrgmod}(x, ...)

\S4method{sodll}{lockedmod}(x, ...)

\S4method{sodll}{packmod}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}
}
\description{
Return the name of the shared object file.
}
\examples{
mod <- mrgsolve:::house()
sodll(mod)

}

