% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_file.R
\name{sort_rbuildignore}
\alias{sort_rbuildignore}
\title{Sort .Rbuildignore file}
\usage{
sort_rbuildignore()
}
\description{
Sorts the .Rbuildignore file using R \code{sort}.
}
\examples{
\dontrun{
sort_rbuildignore()
}
}
