% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spell_check_notebooks}
\alias{spell_check_notebooks}
\title{Spell Check Notebooks}
\usage{
spell_check_notebooks(
  path = "analysis",
  glob = "*.Rmd",
  use_wordlist = TRUE,
  lang = NULL
)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{glob}{A wildcard aka globbing pattern (e.g. \verb{*.csv}) passed on to \code{\link[=grep]{grep()}} to filter paths.}

\item{use_wordlist}{ignore words in the package \link[spelling:wordlist]{WORDLIST} file.}

\item{lang}{set \code{Language} field in \code{DESCRIPTION} e.g. \code{"en-US"} or \code{"en-GB"}.
For supporting other languages, see the \href{https://docs.ropensci.org/hunspell/articles/intro.html#hunspell-dictionaries}{hunspell vignette}.}
}
\description{
Perform a spell check on notebooks with \code{\link[spelling:spell_check_files]{spelling::spell_check_files()}}.
}
\details{
If \code{lang} is \code{NULL} (the default), get language from \code{DESCRIPTION} using
\code{\link[desc:desc_get_field]{desc::desc_get_field()}}.
}
