% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_system.R
\name{sys_info}
\alias{sys_info}
\title{System Information}
\usage{
sys_info()
}
\description{
Simple wrapper around \code{\link[base:Sys.info]{base::Sys.info()}}.
}
\seealso{
\code{\link[base:Sys.info]{base::Sys.info()}}
}
