% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_system.R
\name{sys_which}
\alias{sys_which}
\alias{sys_where}
\title{\code{sys_which} - System \code{which}}
\usage{
sys_which(x, winslash = "/", ...)

sys_where(x, winslash = "/", ...)
}
\arguments{
\item{x}{Passed to \code{Sys.which()} \code{names} argument.}

\item{...}{
  Arguments passed on to \code{\link{normalize_path}}
  \describe{
    \item{\code{path}}{character vector of file paths.}
    \item{\code{winslash}}{the separator to be used on Windows -- ignored
    elsewhere.  Must be one of \code{c("/", "\\\\")}.}
  }}
}
\value{
Character vector of paths, if found. If not found returns \code{NULL} instead of \code{""}.
}
\description{
Lightweight, convenience wrapper around \code{\link[base:Sys.which]{base::Sys.which()}} and \code{\link[base:normalizePath]{base::normalizePath()}}.

Also provided via the alias \code{sys_where()} meant to associate with \code{where.exe} on Windows.
}
\examples{
\dontrun{
sys_which("gdal")
}
}
