% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viridis_quo.R
\name{theme_quo}
\alias{theme_quo}
\title{Quo Theme}
\usage{
theme_quo(
  major = TRUE,
  major.x = TRUE,
  major.y = TRUE,
  minor = TRUE,
  minor.x = TRUE,
  minor.y = TRUE
)
}
\arguments{
\item{major, major.x, major.y, minor, minor.x, minor.y}{set to FALSE to disable grid lines by setting
the appropriate panel.grid element to \code{\link[ggplot2:element]{ggplot2::element_blank()}}.}
}
\description{
An opinionated \link[ggplot2:ggplot2-package]{ggplot2} theme based on \code{\link[ggplot2:ggtheme]{ggplot2::theme_bw()}}. Quo can
be added to individual plots or set as the default theme using \code{\link[=viridis_quo]{viridis_quo()}}.
}
\details{
Quo requires \href{https://www.latofonts.com}{Lato}, which can be installed on macOS using
\verb{brew install font-lato}.
}
\examples{
# adapted from ggplot2::theme_bw()
library(ggplot2)
library(showtext)

font_add_google("Lato", "Lato")
showtext_auto()
# workaround for https://github.com/yixuan/showtext/issues/51
showtext_opts(dpi = 192)

ggplot(mtcars, aes(x = wt, y = mpg, color = factor(cyl))) +
  geom_point() +
  labs(
    title = "Fuel economy declines as weight increases",
    subtitle = "(1973-74)",
    caption = "Data from the 1974 Motor Trend US magazine.",
    tag = "Figure 1",
    x = "Weight (1000 lbs)",
    y = "Fuel economy (mpg)",
    color = "Gears"
  ) +
  theme_quo() +
  scale_color_viridis_d()
}
