% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve.R
\name{touch_funs}
\alias{touch_funs}
\title{Get inits from compiled function.}
\usage{
touch_funs(x, funp = NULL)
}
\arguments{
\item{x}{mrgmod model object}

\item{funp}{optional list of model function pointers}
}
\description{
Get inits from compiled function.
}

