% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimhelp.R
\name{trans}
\alias{trans}
\title{Transform parameter values.}
\usage{
trans(p, x)
}
\arguments{
\item{p}{pars object}

\item{x}{optinal values vector of values with same length as p to transform}
}
\description{
Transform parameter values.
}

