% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{tscale}
\alias{tscale}
\title{Set the \code{tscale} argument for \code{mrgsim}.}
\usage{
tscale(x, value = 1, ...)
}
\arguments{
\item{x}{model object}

\item{value}{value by which time will be scaled}

\item{...}{passed along}
}
\description{
Set the \code{tscale} argument for \code{mrgsim}.
}

