% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{unfreeze}
\alias{unfreeze}
\title{Unfreeze Quarto site}
\usage{
unfreeze()
}
\description{
Delete the Quarto \verb{_freeze} directory to fully re-render the site when \code{\link[quarto:quarto_render]{quarto::quarto_render()}}
is called.
}
