% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{unloadso}
\alias{unloadso}
\alias{unloadso,mrgmod-method}
\title{Unload the model shared object.}
\usage{
unloadso(x, ...)

\S4method{unloadso}{mrgmod}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}
}
\description{
Unload the model shared object.
}

