% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimhelp.R
\name{untrans}
\alias{untrans}
\title{Untransform parameter values.}
\usage{
untrans(p, x)
}
\arguments{
\item{p}{pars object}

\item{x}{optional values vector with same length as p to untransform}
}
\description{
Untransform parameter values.
}

