% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{update_ppm_snapshot}
\alias{update_ppm_snapshot}
\title{Update Posit Package Manager Snapshot Date}
\usage{
update_ppm_snapshot(cooldown = 7, prompt = rlang::is_interactive())
}
\arguments{
\item{cooldown}{Number of days ago to set the PPM snapshot date.}

\item{prompt}{Boolean; prompt the user before updating \link[renv:lockfiles]{renv::lockfiles}?}
}
\value{
Results of \code{\link[renv:lockfile_write]{renv::lockfile_write()}}, invisibly
}
\description{
Set the renv lockfile CRAN repository URL to a Posit Package Manager (PPM) snapshot date
following PPM set up \href{https://packagemanager.posit.co/client/#/repos/cran/setup}{instructions},
for example, \verb{https://packagemanager.posit.co/cran/2026-08-04}.
}
\details{
Using a PPM snapshot offers two key advantages:
\itemize{
\item Faster restores and greater reproducibility using \code{\link[renv:restore]{renv::restore()}}, since renv will restore
the PPM archived macOS and Windows binaries from that date, which will not be available at
\verb{https://packagemanager.posit.co/cran/latest} if newer versions have been published.
\item The snapshot date can be used to implement a CRAN \href{https://cooldowns.dev}{dependency cooldown},
which delays installing the latest version of packages unless they are at least N days old,
(typically 7), providing some protection against supply chain attacks.
}
}
