% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_wordlist_notebooks}
\alias{update_wordlist_notebooks}
\title{Update WORDLIST from notebooks}
\usage{
update_wordlist_notebooks(
  pkg = ".",
  vignettes = TRUE,
  path = "analysis",
  glob = "*.Rmd",
  confirm = TRUE
)
}
\arguments{
\item{pkg}{path to package root directory containing the \code{DESCRIPTION} file}

\item{vignettes}{check all \code{rmd} and \code{rnw} files in the pkg root directory (e.g.
\code{readme.md}) and package \code{vignettes} folder.}

\item{path}{A character vector of one or more paths.}

\item{glob}{A wildcard aka globbing pattern (e.g. \verb{*.csv}) passed on to \code{\link[=grep]{grep()}} to filter paths.}

\item{confirm}{show changes and ask confirmation before adding new words to the list}
}
\description{
Update package \code{inst/WORDLIST} with words from \code{\link[=spell_check_notebooks]{spell_check_notebooks()}}.
}
\details{
\code{update_wordlist_notebooks} is a customized version of \code{\link[spelling:wordlist]{spelling::update_wordlist()}} using
code from \href{https://github.com/ropensci/spelling/blob/master/R/wordlist.R}{\code{wordlist.R}}.
}
