% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upkeep.R
\name{upkeep_checklist}
\alias{upkeep_checklist}
\title{Create upkeep checklist}
\usage{
upkeep_checklist(last_upkeep = last_upkeep_year())
}
\arguments{
\item{last_upkeep}{Year when upkeep was last performed.}
}
\value{
Upkeep checklist for current year as a GitHub markdown array.
}
\description{
Build an upkeep checklist following the format of \code{\link[usethis:tidyverse]{usethis::use_tidy_upkeep_issue()}}.
}
