% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_analysis_package}
\alias{use_analysis_package}
\title{Use Analysis Package Layout}
\usage{
use_analysis_package(prompt = FALSE)
}
\arguments{
\item{prompt}{If TRUE, prompt before writing \code{renv.lock}, passed to \code{\link[renv:snapshot]{renv::snapshot()}}.}
}
\value{
List containing \code{dirs} created, \code{rbuildignore} lines added to .Rbuildignore, \code{gitignore}
exclusions added to .gitignore.
}
\description{
Add the \href{https://jabenninghoff.github.io/rdev/articles/analysis-package-layout.html}{Analysis Package Layout}
to the current package.
}
\details{
When run, \code{use_analysis_package()} creates analysis package directories, adds exclusions to
.gitignore and .Rbuildignore, creates \verb{_base.yml} in \code{pkgdown} from the first \code{URL} in
\code{DESCRIPTION}, installs the \code{README.Rmd} template for analysis packages, and the \code{dplyr}
package needed for the \code{README.Rmd} template.
}
