% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_codecov}
\alias{use_codecov}
\title{Use rdev codecov}
\usage{
use_codecov()
}
\description{
Install code coverage with \code{\link[usethis:use_coverage]{usethis::use_coverage(type = "codecov")}},
\code{DT} package for \code{\link[covr:report]{covr::report()}}, and rdev GitHub action \code{test-coverage.yaml}.
}
\details{
Because \code{\link[=use_rdev_package]{use_rdev_package()}}, \code{\link[=use_analysis_package]{use_analysis_package()}} and \code{use_codecov()} all modify README.Rmd,
\code{use_codecov()} must be run last or its changes will be overwritten. \code{use_codecov()} is not run
in \code{\link[=use_rdev_package]{use_rdev_package()}}.
}
