% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_lintr}
\alias{use_lintr}
\title{Use rdev .lintr}
\usage{
use_lintr(open = FALSE)
}
\arguments{
\item{open}{Open the newly created file for editing? Happens in RStudio or
Positron, if applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Install rdev .lintr template using \code{\link[usethis:use_template]{usethis::use_template()}}
}
