% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_rdev_package}
\alias{use_rdev_package}
\title{Use rdev package conventions}
\usage{
use_rdev_package(quiet = TRUE)
}
\arguments{
\item{quiet}{If TRUE, disable user prompts by setting \code{\link[rlang:is_interactive]{rlang::local_interactive()}} to FALSE.}
}
\description{
Add rdev templates and settings within the active package. Normally invoked when first setting
up a package.
}
\section{GitHub Pages}{
 GitHub Pages is enabled by default for public repositories, and can be
disabled by setting \code{rdev.github.pages} to \code{FALSE}: \code{options(rdev.github.pages = FALSE)}.
GitHub Pages is disabled by default for private repositories (as it is not supported on the
free plan), and can be enabled by setting \code{rdev.github.pages} to \code{TRUE}.
}

\section{GitHub Actions}{
 GitHub Actions can be disabled by setting \code{rdev.github.actions} to
\code{FALSE}: \code{options(rdev.github.actions = FALSE)}
}

