% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rdev.R
\name{use_rdev_rprofile}
\alias{use_rdev_rprofile}
\title{Use rdev .Rprofile}
\usage{
use_rdev_rprofile(directory = ".")
}
\arguments{
\item{directory}{Directory relative to active project to update .Rprofile}
}
\description{
Update .Rprofile to attach devtools and rdev when in an interactive session.
}
\details{
\strong{Warning:} \code{use_rdev_rprofile()} won't work properly if some (but not all) of the added
lines are already present in .Rprofile
}
