% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_spelling}
\alias{use_spelling}
\title{Use rdev spelling}
\usage{
use_spelling(lang = "en-US", prompt = FALSE)
}
\arguments{
\item{lang}{Preferred spelling language. Usually either \code{"en-US"} or
\code{"en-GB"}.}

\item{prompt}{If TRUE, prompt before writing \code{renv.lock}, passed to \code{\link[renv:snapshot]{renv::snapshot()}}.}
}
\description{
Install \link[spelling:spell_check_package]{spelling} with rdev conventions.
}
\details{
Since \code{\link[spelling:spell_check_package]{spelling::spell_check_setup()}} requires user interaction, \code{use_spelling()} is not run in
\code{\link[=use_rdev_package]{use_rdev_package()}}.
}
