% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_todo}
\alias{use_todo}
\title{Use rdev TODO.md}
\usage{
use_todo(open = rlang::is_interactive())
}
\arguments{
\item{open}{Open the newly created file for editing? Happens in RStudio or
Positron, if applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Install rdev TODO.md template using \code{\link[usethis:use_template]{usethis::use_template()}}
}
