% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upkeep.R
\name{use_upkeep_issue}
\alias{use_upkeep_issue}
\title{Create an upkeep checklist in a GitHub issue}
\usage{
use_upkeep_issue(last_upkeep = last_upkeep_year())
}
\arguments{
\item{last_upkeep}{Year of last upkeep. By default, the \code{Config/rdev/last-upkeep} field in
\code{DESCRIPTION} is consulted for this, if it's defined. If there's no information on the last
upkeep, the issue will contain the full checklist.}
}
\description{
Opens an issue in the package repository with a checklist of tasks for regular maintenance and
adherence to rdev package standards. Some tasks are meant to be done only once, and others should
be reviewed periodically. Adapted from \code{\link[usethis:use_upkeep_issue]{usethis::use_upkeep_issue()}} and
\code{\link[usethis:tidyverse]{usethis::use_tidy_upkeep_issue()}}.
}
