% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cmd.R
\name{utils_cmd}
\alias{utils_cmd}
\alias{cmd}
\alias{git}
\alias{gh}
\alias{docker}
\alias{npx}
\title{Command Utilities}
\usage{
cmd(cmd, ..., .timeout = 60, .echo = TRUE)

git(cmd, ..., .timeout = 60, .echo = TRUE)

gh(cmd, ..., .timeout = 60, .echo = TRUE)

docker(cmd, ..., .timeout = 60, .echo = TRUE)

npx(cmd, ..., .timeout = 60, .echo = TRUE)
}
\arguments{
\item{cmd}{The command to run.}

\item{...}{Additional arguments to pass to the command.}

\item{.timeout}{The timeout in seconds for the command to complete. Default is 60 seconds.}

\item{.echo}{Logical. If \code{TRUE}, the output will be printed to the console. Default is \code{TRUE}.}
}
\value{
The output of the command.
}
\description{
Utilities for running system commands and tools like \code{git}, \code{docker}, etc.

These functions leverage \code{\link[sys:exec_internal]{sys::exec_internal()}} to run commands and capture their output.
\itemize{
\item \code{cmd()}: Run a generic command.
\item \code{git()}: Run a \code{git} command.
\item \code{gh()}: Run a GitHub CLI (\code{gh}) command.
\item \code{docker()}: Run a \code{docker} command.
}
}
\examples{
#--- git ---
\dontrun{
# get git version
git("--version")
# git status
git("status")
# git commit
git("commit", "-m", "Initial commit")
}

#--- gh ---
\dontrun{
# command used to initialize this repository on github:
gh("repo", "create", "landrise/landrise.utils", "--private", "--source=.", "--remote=upstream")
}

# --- docker ---
\dontrun{
# get docker help
docker("--help")
# list docker images
docker("images")
# run a docker container
docker("run", "--rm", "hello-world")
}

# --- npx ---
\dontrun{
# get npx help
npx("--help")
# create a new react app
npx("create-react-app", "my-app")
# run repomix tool
npx("repomix", "--help")
}
}
