% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsims.R
\docType{methods}
\name{variables}
\alias{variables}
\alias{variables,mrgsims-method}
\title{Return the requested compartments and tabled items from a simulation run.}
\usage{
variables(x, ...)

\S4method{variables}{mrgsims}(x, ...)
}
\arguments{
\item{x}{mrgsims object}

\item{...}{passed along}
}
\description{
Return the requested compartments and tabled items from a simulation run.
}
\seealso{
\code{\link{request}}
}

