% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viridis_quo.R
\name{viridis_quo}
\alias{viridis_quo}
\title{Viridis Quo}
\usage{
viridis_quo()
}
\description{
Sets the default theme to \code{\link[=theme_quo]{theme_quo()}} and the default color scales to
\code{\link[ggplot2:scale_viridis]{viridis}}.
}
\details{
To reset the default theme and color scales, restart the R session.
}
\examples{
\dontrun{

viridis_quo()
}
}
