% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_eval}
\alias{write_eval}
\title{Write and evaluate an expression}
\usage{
write_eval(string)
}
\arguments{
\item{string}{An expression to be printed to the console and evaluated}
}
\value{
The return value from the evaluated expression
}
\description{
\code{write_eval(string)} is a simple wrapper that prints \code{string} to the console using
\code{\link[base:writeLines]{writeLines()}}, then executes the expression using \code{\link[base:parse]{parse()}}
and \code{\link[base:eval]{eval()}}.
}
\examples{
write_eval("pi")

write_eval("exp(1)")
}
