% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_rdev_site}
\alias{build_rdev_site}
\title{Build rdev Site}
\usage{
build_rdev_site(pkg = ".", ...)
}
\arguments{
\item{pkg}{Path to package. Currently, only \code{pkg = "."} is supported.}

\item{...}{additional arguments passed to \code{\link[pkgdown:build_site]{pkgdown::build_site()}} (not implemented)}
}
\description{
\code{build_rdev_site()} is a wrapper for \code{\link[pkgdown:build_site]{pkgdown::build_site()}} optimized for rdev workflow that
updates \code{README.md} and performs a clean build using \code{pkgdown}.
}
\details{
When run, \code{build_rdev_site()} calls:
\enumerate{
\item \code{\link[devtools:build_rmd]{devtools::build_readme()}}
\item \code{\link[pkgdown:clean_site]{pkgdown::clean_site()}}
\item \code{\link[pkgdown:build_site]{pkgdown::build_site()}}
}
}
\examples{
\dontrun{
build_rdev_site()
}
}
