% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{carry_out}
\alias{carry_out}
\title{Set the \code{carry.out} argument for \code{mrgsim}.}
\usage{
carry_out(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}
}
\description{
Set the \code{carry.out} argument for \code{mrgsim}.
}
\examples{
data(extran3)

mod <- mrgsolve:::house()

head(extran3)

mod \%>\% 
  data_set(extran3, ID == 1) \%>\% 
  carry_out(CL,KA,amt,ii) \%>\% 
  mrgsim

}

