% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_renv.R
\name{check_renv}
\alias{check_renv}
\title{Check renv}
\usage{
check_renv(update = TRUE)
}
\arguments{
\item{update}{run \code{renv::update} (default \code{TRUE})}
}
\description{
Checks \code{renv} \code{status()}, \code{clean()}, and optionally \code{update()}
}
\examples{
\dontrun{
check_renv()
check_renv(update = FALSE)
}
}
