% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\title{Local CI}
\usage{
ci(renv = TRUE, styler = NULL, lintr = TRUE, document = TRUE, rcmdcheck = TRUE)
}
\arguments{
\item{renv}{check \code{\link[renv:status]{renv::status()}}}

\item{styler}{style all files using \code{\link[=style_all]{style_all()}}, see details}

\item{lintr}{lint all files using \code{\link[=lint_all]{lint_all()}}}

\item{document}{run \code{\link[devtools:document]{devtools::document()}}}

\item{rcmdcheck}{run \code{R CMD check} using:
\code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck(args = "--no-manual", error_on = "warning")}}}
}
\description{
Run continuous integration tests locally.
}
\details{
If \code{\link[renv:status]{renv::status()}} is not synchronized, \code{ci()} will stop.

If \code{styler} is set to \code{NULL} (the default), \code{\link[=style_all]{style_all()}} will be run only if there are no
uncommitted changes to git. Setting the value to \code{TRUE} or \code{FALSE} overrides this check.

If \code{\link[=lint_all]{lint_all()}} finds any lints, \code{ci()} will stop and open the RStudio markers pane.
}
\examples{
\dontrun{
ci()
ci(styler = TRUE)
ci(styler = FALSE, rcmdcheck = FALSE)
}
}
