% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\title{Local CI}
\usage{
ci(styler = FALSE, lintr = FALSE, document = TRUE, rcmdcheck = TRUE)
}
\arguments{
\item{styler}{style all files using \code{\link[=style_all]{style_all()}}}

\item{lintr}{lint all files using \code{\link[=lint_all]{lint_all()}}}

\item{document}{run \code{\link[devtools:document]{devtools::document()}}}

\item{rcmdcheck}{run \code{R CMD check} using:
\code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck(args = "--no-manual", error_on = "warning")}}}
}
\description{
Run continuous integration tests locally.
}
\examples{
\dontrun{
ci()
ci(styler = TRUE)
ci(styler = TRUE, lintr = TRUE, rcmdcheck = FALSE)
}
}
