% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\title{Local CI}
\usage{
ci(styler = FALSE, lintr = TRUE, rcmdcheck = TRUE)
}
\arguments{
\item{styler}{style all files: \code{styler::style_dir(".", exclude_dirs = c("renv"))}}

\item{lintr}{lint all files:
\code{lintr::lint_dir(path = ".", exclusions = list(".git", ".github", "packrat", "renv"))}}

\item{rcmdcheck}{run \code{R CMD check} using
\code{rcmdcheck::rcmdcheck(args = "--no-manual", error_on = "error")}}
}
\description{
Run continuous integration tests locally.
}
\examples{
\dontrun{
ci()
ci(styler = TRUE)
ci(styler = TRUE, lintr = FALSE, rcmdcheck = FALSE)
}
}
