% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{FIXED}
\alias{FIXED}
\title{Parse \code{$FIXED} block.}
\usage{
FIXED(x, env, annotated = FALSE, pos = 1, object = NULL, ...)
}
\arguments{
\item{x}{data}

\item{env}{parse environment}

\item{annotated}{logical}

\item{pos}{parse position}

\item{object}{the (character) name of a \code{list} in \code{$ENV} to use for block output}

\item{...}{passed}
}
\description{
Parse \code{$FIXED} block.
}
\details{
If \code{object} is specified, it must be the name of a named list.
}

