% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlchecker.R
\name{html_url_check}
\alias{html_url_check}
\title{Check URLs in HTML files}
\usage{
html_url_check(
  path = "docs",
  parallel = TRUE,
  pool = curl::new_pool(),
  progress = TRUE
)
}
\arguments{
\item{path}{Path to the directory of HTML files}

\item{parallel}{If \code{TRUE}, check the URLs in parallel}

\item{pool}{A multi handle created by \code{\link[curl:new_pool]{curl::new_pool()}}. If \code{NULL} use a global pool.}

\item{progress}{Whether to show the progress bar for parallel checks}
}
\value{
A \code{url_checker_db} object (invisibly). This is a \code{check_url_db} object with an added
class with a custom print method.
}
\description{
Runs \code{\link[urlchecker:url_check]{urlchecker::url_check()}} with a database created using the \code{url_db_from_HTML_files}
function in the tools package.
}
