% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{lint_all}
\alias{lint_all}
\title{Lint all files}
\usage{
lint_all(path = ".", ...)
}
\arguments{
\item{path}{For the base directory of the project (for \code{lint_dir()}) or
package (for \code{lint_package()}).}

\item{...}{
  Arguments passed on to \code{\link[lintr:lint]{lintr::lint_dir}}
  \describe{
    \item{\code{parse_settings}}{whether to try and parse the settings.}
    \item{\code{relative_path}}{if \code{TRUE}, file paths are printed using their path relative to the base directory.
If \code{FALSE}, use the full absolute path.}
    \item{\code{exclusions}}{exclusions for \code{\link[lintr:exclude]{exclude()}}, relative to the package path.}
    \item{\code{pattern}}{pattern for files, by default it will take files with any of the extensions .R, .Rmd, .Rnw, .Rhtml,
.Rrst, .Rtex, .Rtxt allowing for lowercase r (.r, ...)}
  }}
}
\value{
A list of lint objects.
}
\description{
Lint all files in a project. Implemented as a wrapper for \code{\link[lintr:lint]{lintr::lint_dir()}}.
}
\examples{
\dontrun{
lint_all()
lint_all("analysis")
}
}
