% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimhelp.R
\name{logit_par}
\alias{logit_par}
\title{Create a list of logit-transformed parameters.}
\usage{
logit_par(name, value, fixed = FALSE, ...)
}
\arguments{
\item{name}{the parameter name}

\item{value}{the parameter value}
}
\description{
Create a list of logit-transformed parameters.
}

