% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintern.R
\name{mintern}
\alias{mintern}
\title{Load a library model.}
\usage{
mintern(model, tryload = TRUE, ...)
}
\arguments{
\item{model}{model name}

\item{tryload}{try to load the model?}

\item{...}{passed to update}
}
\description{
Load a library model.
}
\examples{
mod <- mread("pk1cmt",mrgmod())
mod <- mread("pk2cmt",mrgmod()) 
mod <- mread("pk3cmt",mrgmod()) 
mod <- mread("irm1",mrgmod()) 
mod <- mread("irm2",mrgmod()) 
mod <- mread("irm3",mrgmod()) 
mod <- mread("irm4",mrgmod())
mod <- mread("emax",mrgmod())
mod <- mread("effect",mrgmod())
mod <- mread("tmdd",mrgmod()) 

}

