% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintern.R
\name{mintern}
\alias{mintern}
\title{Load a library model.}
\usage{
mintern(model, tryload = TRUE, ...)
}
\arguments{
\item{model}{model name}

\item{tryload}{try to load the model?}

\item{...}{passed to update}
}
\description{
Load a library model.
}
\examples{
mod <- mread("pk1cmt",modlib())
mod <- mread("pk2cmt",modlib()) 
mod <- mread("pk3cmt",modlib()) 
mod <- mread("irm1",modlib()) 
mod <- mread("irm2",modlib()) 
mod <- mread("irm3",modlib()) 
mod <- mread("irm4",modlib())
mod <- mread("emax",modlib())
mod <- mread("effect",modlib())
mod <- mread("tmdd",modlib()) 

}

