% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintern.R
\name{mintern}
\alias{mintern}
\title{Load a library model.}
\usage{
mintern(model, tryload = TRUE, ...)
}
\arguments{
\item{model}{model name}

\item{tryload}{try to load the model?}

\item{...}{passed to update}
}
\description{
Load a library model.
}
\examples{
mod <- mintern("pk1cmt")
mod <- mintern("pk2cmt") 
mod <- mintern("pk3cmt") 
mod <- mintern("irm1") 
mod <- mintern("irm2") 
mod <- mintern("irm3") 
mod <- mintern("irm4")
mod <- mintern("emax")
mod <- mintern("effect")
mod <- mintern("tmdd") 

}

