% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintern.R
\name{mintern}
\alias{mintern}
\title{Load a library model.}
\usage{
mintern(model, tryload = TRUE, ...)
}
\arguments{
\item{model}{model name}

\item{tryload}{try to load the model?}

\item{...}{passed to update}
}
\description{
Load a library model.
}
\examples{
mod <- mread("pk1cmt",mlib())
mod <- mread("pk2cmt",mlib()) 
mod <- mread("pk3cmt",mlib()) 
mod <- mread("irm1",mlib()) 
mod <- mread("irm2",mlib()) 
mod <- mread("irm3",mlib()) 
mod <- mread("irm4",mlib())
mod <- mread("emax",mlib())
mod <- mread("effect",mlib())
mod <- mread("tmdd",mlib()) 

}

