% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_type}
\alias{package_type}
\title{Determine rdev package type}
\usage{
package_type(pkg = ".", strict = FALSE)
}
\arguments{
\item{pkg}{path to package}

\item{strict}{strictly determine package type (see description)}
}
\value{
type string, one of \code{c("rdev", "analysis", "quarto")}
}
\description{
Use heuristics (when \code{strict = FALSE}) or presence of specific files (when \code{strict = TRUE}) to
determine package type. If one of the files is found, the package is strictly determined:
}
\details{
\itemize{
\item If \verb{_quarto.yml} is found, the type is \code{quarto}
\item If \verb{pkgdown/_base.yml} is found, the type is \code{analysis}
\item If \verb{_pkgdown.yml} is found, the type is \code{rdev}
}

If strict checking fails, the package is assumed to be \code{analysis} if an \code{analysis} directory is
present, and \code{rdev} if not.
}
