% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rdev-package}
\alias{rdev}
\alias{rdev-package}
\title{rdev: R Development Tools}
\description{
An opinionated collection of development tools, packages, and utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jabenninghoff.github.io/rdev}
  \item \url{https://github.com/jabenninghoff/rdev}
  \item Report bugs at \url{https://github.com/jabenninghoff/rdev/issues}
}

}
\author{
\strong{Maintainer}: John Benninghoff \email{jbenninghoff@mac.com} (\href{https://orcid.org/0000-0002-6230-4742}{ORCID})

}
\keyword{internal}
