% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\docType{methods}
\name{render}
\alias{render}
\alias{render,character-method}
\alias{render,mrgmod-method}
\title{Render an \code{mrgsolve} model.}
\usage{
render(x, ...)

\S4method{render}{character}(x, project, ...)

\S4method{render}{mrgmod}(x, ...)
}
\arguments{
\item{x}{model object or the model name}

\item{...}{passed to \code{rmarkdown::render}}

\item{project}{the directory containing the \code{.cpp} model file}
}
\description{
Render an \code{mrgsolve} model.
}
\examples{
\dontrun{
mod <- mrgsolve:::house()
mrgsolve:::render(mod)
mrgsolve:::render("irm2", modlib())
}

}

