% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{setup_analysis}
\alias{setup_analysis}
\title{Set up analysis package}
\usage{
setup_analysis()
}
\description{
Set up an analysis package within an rdev package newly initialized with \code{\link[=init]{init()}}, by committing
and running \code{\link[=use_analysis_package]{use_analysis_package()}}, \code{\link[=use_spelling]{use_spelling()}}, and \code{\link[=ci]{ci()}}.
}
\details{
\code{setup_analysis()} will stop if \code{\link[rlang:is_interactive]{rlang::is_interactive()}} is \code{FALSE}.
}
