% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_file.R
\name{sort_file}
\alias{sort_file}
\title{Sort file}
\usage{
sort_file(name)
}
\arguments{
\item{name}{file to be sorted}
}
\description{
Sorts a file using R \code{\link[=sort]{sort()}}.
}
\examples{
\dontrun{
sort_file(".Rbuildignore")
}
}
\seealso{
\code{\link[=sort_rbuildignore]{sort_rbuildignore()}}
}
