% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{stage_release}
\alias{stage_release}
\title{Stage a GitHub release}
\usage{
stage_release(pkg = ".", filename = "NEWS.md")
}
\arguments{
\item{pkg}{path to package. Currently, only \code{pkg = "."} is supported.}

\item{filename}{name of file containing release notes, defaults to \code{NEWS.md}.}
}
\description{
Open a GitHub pull request for a new release from \code{NEWS.md}. Approve, merge, and create the
release using \code{merge_release()}.
}
\details{
When run, \code{stage_release()}:
\enumerate{
\item Extracts release version and release notes from \code{NEWS.md} using \code{\link[=get_release]{get_release()}}
\item Validates version conforms to rdev conventions (#.#.#) and release notes aren't empty
\item Verifies that version tag doesn't already exist using \code{\link[gert:git_tag]{gert::git_tag_list()}}
\item Checks for uncommitted changes and stops if any exist using \code{\link[gert:git_diff]{gert::git_diff_patch()}}
\item Creates a new branch if on the default branch (\code{\link[gert:git_branch]{gert::git_branch()}} \code{==}
\code{\link[usethis:git-default-branch]{usethis::git_default_branch()}}) using \code{\link[gert:git_branch]{gert::git_branch_create()}}
}
}
