% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_all.R
\name{style_all}
\alias{style_all}
\title{Style all files}
\usage{
style_all(
  path = ".",
  filetype = c("R", "Rprofile", "Rmd"),
  exclude_dirs = c(".git", ".github", "packrat", "renv")
)
}
\arguments{
\item{path}{Path to a directory with files to transform.}

\item{filetype}{Vector of file extensions indicating which file types should
be styled. Case is ignored, and the \code{.} is optional, e.g.
\code{c(".R", ".Rmd")}, or \code{c("r", "rmd")}. Supported values (after
standardization) are: "r", "rprofile", "rmd", "rnw".}

\item{exclude_dirs}{Character vector with directories to exclude
(recursively).}
}
\description{
Style all files in a project. Implemented as a wrapper for styler::style_dir() that styles .R and
.Rmd filetypes, and excludes .git, .github, packrat, and renv by default.
}
\examples{
\dontrun{
style_all()
style_all("notebooks")
style_all(
  path = "./inst",
  filetype = "r",
  exclude_dirs = c("inst/example/bad.R")
)
}
}
