% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_document.R
\name{to_document}
\alias{to_document}
\title{Convert R Notebook to \code{html_document}}
\usage{
to_document(file_path, new_path, overwrite = FALSE)
}
\arguments{
\item{file_path}{Path to the source file}

\item{new_path}{Path to copy the converted file using \code{\link[fs:copy]{fs::file_copy()}}}

\item{overwrite}{Overwrite file if it exists, passed to \code{\link[fs:copy]{fs::file_copy()}}}
}
\value{
Path to new file
}
\description{
Copies a file using \code{\link[fs:copy]{fs::file_copy()}} and changes the output type in the yaml front matter from
\code{html_notebook} to \code{html_document}, removing all other output types.
}
\examples{
\dontrun{
to_document("notebook.Rmd", "document.Rmd")
to_document("notebooks_dir/notebook.Rmd", "documents_dir")
}
}
\seealso{
\code{\link[=build_analysis_site]{build_analysis_site()}}
}
