% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ranalysis.R
\name{use_analysis_package}
\alias{use_analysis_package}
\title{Use Analysis Package Layout}
\usage{
use_analysis_package()
}
\description{
Add the \href{https://jabenninghoff.github.io/rdev/articles/analysis-package-layout.html}{Analysis Package Layout}
to the current package.
}
\details{
When run, \code{use_analysis_package()} creates analysis package directories and updates both
.gitignore and .Rbuildignore.
}
\examples{
\dontrun{
use_analysis_package()
}
}
