% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_gitattributes}
\alias{use_gitattributes}
\title{Use rdev .gitattributes}
\usage{
use_gitattributes(open = FALSE)
}
\arguments{
\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Install rdev .gitattributes template using \code{\link[usethis:use_template]{usethis::use_template()}} to set GitHub Linguist
\href{https://github.com/github-linguist/linguist/blob/main/docs/overrides.md}{overrides}. If an
analysis package is detected, RMarkdown will be enabled as a language for GitHub statistics.
}
