% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rdev.R
\name{use_lintr}
\alias{use_lintr}
\title{Use rdev .lintr}
\usage{
use_lintr(directory = ".")
}
\arguments{
\item{directory}{Directory relative to active project to update .lintr}
}
\description{
Add .lintr file with 'linters: with_defaults(line_length_linter(100))' to current project
}
