% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_package_r}
\alias{use_package_r}
\title{Use rdev package.R}
\usage{
use_package_r(open = FALSE)
}
\arguments{
\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Install rdev package.R template using \code{\link[usethis:use_template]{usethis::use_template()}}
}
\details{
package.R is saved as "R/package.R".
}
