% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_rdev_package}
\alias{use_rdev_package}
\title{Use rdev package conventions}
\usage{
use_rdev_package(quiet = TRUE)
}
\arguments{
\item{quiet}{If TRUE, disable user prompts and suppress output when testing.}
}
\description{
Add rdev templates and settings within the active package. Normally invoked when first setting
up a package.
}
