% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rdev.R
\name{use_rprofile}
\alias{use_rprofile}
\title{Use rdev .Rprofile}
\usage{
use_rprofile(open = FALSE)
}
\arguments{
\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Install rdev .Rprofile template using \code{\link[usethis:use_template]{usethis::use_template()}}
}
